Can Load in Own Images
You can import textures into it now. The tilebutton pictureboxes also get created through code. The problem that this has caused is that when i save a file that has used custom textures. When loaded in, if there isn't enough in the list then it will break. If the same assets have been imported before loading in, it DOES work.